% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meuse.riv}
\alias{meuse.riv}
\title{River Meuse outline}
\format{This data frame polygon of the Meuse river in a latitude longitude format.}
\source{
See the \code{\link{meuse.riv}} for more documentation.
}
\usage{
data(meuse.riv)
}
\description{
The meuse.riv data consists of an outline of the Meuse
river in the area a few kilometers around the \code{\link{meuse}} data set.
}
\examples{
 \dontrun{
data(meuse.riv)
spRplot(meuse.riv,col.pallete=list(col='blue'))
}
}
\keyword{datasets}

