% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{+.spGoogle}
\alias{+.spGoogle}
\title{Modify a spGoogle object by adding on new components.}
\usage{
\method{+}{spGoogle}(e1, e2, ...)
}
\arguments{
\item{e1}{object of the spGoogle class}

\item{e2}{object of the spGoogle class}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This operator allows you to add objects to a spGoogle object.
}

