% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.spGoogle}
\alias{print.spGoogle}
\title{Invisible print to spGoogle objects.}
\usage{
\method{print}{spGoogle}(x, ...)
}
\arguments{
\item{x}{object of the spGoogle class}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Invisibly returns the result of spGplot 
which is a list with components that contain the kml and legend path.
}

