% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spGplot.R
\name{spGplot}
\alias{spGplot}
\title{Export R information to Google Maps}
\usage{
spGplot(data, var = NULL, maptype = "satellite", description = list(var =
  NULL, type = NULL, title = NULL), decimals = 3, col.pallete = list(col =
  heat.colors(if (length(cuts) == 1) cuts else length(cuts)), alpha = 1),
  legend.att = list(title = NULL, ncol = 1, bg = "#FFFFFF"), cuts = 5,
  cuts.type = "range", lwd = 1.5, sizeBall = c(0.5, 1), savekml = NULL,
  map.name = "kml", map.description = "description", google.maps = TRUE,
  google.earth.path = try(system("which google-earth", ignore.stdout = T,
  ignore.stderr = T), silent = T), ...)
}
\arguments{
\item{data}{a spatial data from any sp classes.}

\item{var}{an attribute name in the spatial data.}

\item{maptype}{GoogleMaps types: 'satellite', 'terrain', 'roadmap', 'satellite', 'hybrid'.}

\item{description}{a \code{\link[base]{list}} containing the description information in the balloon. var = attribute 
name(s) in the spatial data. type = type of plot inside the balloon: 'table', 'bar', 'pie' and 'png'. title = balloon title.}

\item{decimals}{number of decimals in the legend.}

\item{col.pallete}{a list containing the pallete of colors to be used (col) and the 
transparency level (alpha) .}

\item{legend.att}{a list containing the attributes for legend. See \code{\link[graphics]{legend}}.}

\item{cuts}{number of cuts in the legend. Can also be a vector with the cuts point.}

\item{cuts.type}{method of selecting the cuts points. Ignored if cuts = vector .}

\item{lwd}{the line width for SpatialLines/SpatialLinesDataframe class.}

\item{sizeBall}{numeric vector of the length 2 providing the min and max of the balls.}

\item{savekml}{the name of the output KML. If name is not defined the function will
create the file in directory temporary. Default is NULL.}

\item{map.name}{the name of the KML.}

\item{map.description}{the description of th KML.}

\item{google.maps}{boolean. Default TRUE. Define if the data should be exported to GoogleMaps.}

\item{google.earth.path}{if google.maps = FALSE then GoogleEarth path must be provided}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A plot window on the browser.
}
\description{
Export to GoogleMaps R spatial data information
}
\examples{
\dontrun{
library(RColorBrewer)
library(spGoogle)
data(rio)

brks <- hist(rio$Dengue, plot = FALSE)$breaks
n_cuts <- length(brks)
pallete <- brewer.pal(n_cuts-1, 'Blues')
names(rio@data)
spGplot(rio, var = "Dengue", cuts = brks, 
        col.pallete = list(col = pallete, alpha = 0.8), 
        description = list(title = 'Name', type = 'table', var = c('Income','Urban')),
        maptype = 'roadmap')
                        
spGplot(rio, var = "Dengue", cuts = brks, 
        col.pallete = list(col = pallete, alpha = 0.8), 
        description = list(title = 'Name', type = 'bar', var = c('Income','Urban')),
        maptype = 'roadmap')
        
spGplot(rio, var = "Dengue", cuts = brks, 
       col.pallete = list(col = pallete, alpha = 0.8), 
       description = list(title = 'Name', type = 'pie', var = c('Income','Urban')),
       maptype = 'roadmap')
       
#custom legend
spGplot(rio, var = "Dengue", cuts = brks, 
       col.pallete = list(col = pallete, alpha = 0.8), 
       description = list(title = 'Name', type = 'pie', var = c('Income','Urban')),
       legend.att = list(title = 'Dengue', bg = 'red', cex = 0.8, ncol = 2)
       maptype = 'roadmap')
}

}
\seealso{
See \code{\link{spRplot}}.
}

