% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meuse.grid}
\alias{meuse.grid}
\title{Prediction Grid for Meuse Data Set}
\format{This data frame contains the following columns:
\describe{
\item{x}{a numeric vector; latitude coordinate (see \code{\link{meuse})} }
\item{y}{a numeric vector; longitude coordinate (see \code{\link{meuse})}}
}
 
 @details \code{x} and \code{y} are in latitude and longitude format. 
 Using +datum=WGS84.
 
 @source \url{http://www.gstat.org/}
 @references See the \code{\link{meuse}} documentation.}
\usage{
data(meuse.grid)
}
\description{
The meuse.grid has a grid with 0.0002 degrees x 0.0002 degrees 
spacing that covers the Meuse study area (see \code{\link{meuse}}) in a latitude longitude format .
}
\examples{
\dontrun{
 data(meuse.grid)
 spRplot(meuse.grid, col.pallete=list(col='lightgrey',alpha=0.5),
 maptype = 'hybrid')
}
}
\keyword{datasets}

